Keep widening tests with core unit tests
